pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "compox"
dynamic = ["version"]
description = ""
authors = [
{name = "Jan Matula", email = "jan.matula@tescan.com"},
{name = "Oldřich Kodym", email = "oldrich.kodym@tescan.com"}
]
dependencies = [
    "fastapi>=0.100.1",
    "boto3>=1.36.0",
    "celery==5.3.1",
    "uvicorn>=0.25.0",
    "h5py>=3.9.0",
    "numpy<2.0.0",
    "natsort==8.4.0",
    "pytest>=7.4.2",
    "pydantic-settings>=2.0.3",
    "python-minifier>=2.9.0",
    "pystray>=0.19.5",
    "pywin32>=306; platform_system=='Windows'",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
    "flower>=2.0.1",
    "python-multipart>=0.0.20",
    "mangum>=0.19.0",
    "requests==2.32.3",
    "pycurl>=7.45.4",
    "loguru>=0.7.3",
    "typer>=0.16.0",
    "pip>=25.2",
    "build>=1.3.0",
    "pdm-backend>=2.4.5",
]
requires-python = ">=3.10.0"
readme = "README.md"
license = "Apache-2.0 AND LicenseRef-Commons-Clause-1.0"
license-files = [
  "LICENSE",
  "NOTICE",
  "THIRD_PARTY_LICENSES.txt"
]

[tool.pdm.version]
source = "scm"
tag_filter = "compox-v*"  # version is parsed using default regex
version_format = "src.compox.version:format_version"
fallback_version = "0.0.0"

[tool.uv]
package = true

[tool.uv.workspace]
members = ["dist"]

[dependency-groups]
docs = [
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]

[project.scripts]
compox = "compox.cli:app"
